# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyHapaDocument",
        "version": "1.2"
    },
    "paths": {
...
        "/api/v1/auth/otp/send": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send OTP",
                "description": "Send OTP to user via SMS",
                "operationId": "sendOtp",
                "requestBody": {
                    "description": "User contact information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "identity",
                                    "channel"
                                ],
                                "properties": {
                                    "identity": {
                                        "description": "User's phone number",
                                        "type": "string",
                                        "example": "+989123456789"
                                    },
                                    "channel": {
                                        "description": "Channel to send OTP",
                                        "type": "string",
                                        "enum": [
                                            "sms"
                                        ],
                                        "example": "sms"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Operation status",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "token": {
                                            "description": "Unique token for OTP verification",
                                            "type": "string",
                                            "example": "667c7641-1921-42ea-966c-cb7766bdfc86"
                                        },
                                        "message": {
                                            "description": "Response message",
                                            "type": "string",
                                            "example": "OTP sent."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
...
        "/api/v1/auth/otp/verify-identity": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify Identity",
                "description": "Verify user's identity using OTP and national code",
                "operationId": "verifyIdentity",
                "requestBody": {
                    "description": "Identity verification data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "birth_date",
                                    "national_code"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "Unique token for OTP verification",
                                        "type": "string",
                                        "example": "667c7641-1921-42ea-966c-cb7766bdfc86"
                                    },
                                    "birth_date": {
                                        "description": "User's birth date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "1310/06/30"
                                    },
                                    "national_code": {
                                        "description": "User's national code",
                                        "type": "string",
                                        "example": "0012345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Operation status",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Response message",
                                            "type": "string",
                                            "example": "Registration and identity verification completed successfully."
                                        },
                                        "user": {
                                            "type": "object"
                                        },
                                        "auth_token": {
                                            "description": "Authentication token",
                                            "type": "string",
                                            "example": "2|iKaaaaaaaaaaaaaaaaaaaa6cU1qr5m2SCBk1TyWd934ae54c"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}